first migration

+++ APINetCore/src/ApiNetCore.Business/Models/Entity.cs
@@ -2,6 +2,6 @@ namespace ApiNetCore.Business.Models
 {
     public class Entity
     {
-        public int Id { get; set; }
+        public ushort Id { get; set; }
     }
 }
